Set lastname not to be required by HTML validations
The address generated by the test factories has no lastname,
and, as is now, it is not possible to go head with checkout specs
filling the address with user data. On a side note, we have to get
rid of the lastname attribute from this gem, in order to be up to
date with Solidus upstream.

Also, the lastname is actually not required in the model validation,
so it would have been the right thing to do anyway.