Invert dependency of SolidusStarterFrontend and Solidus assets
Goal
----

As a SolidusStarterFrontend contributor

I want to invert the dependencies of SolidusStarterFrontend and Solidus
assets such that the SolidusStarterFrontend manifests depend on the
Solidus assets (and not the other way around)

So that we can simplify the frontends generated by the gem

Background
----------

From
https://github.com/nebulab/solidus_starter_frontend/pull/174#issuecomment-896682012:

> Can we take a step back to understand better why this is needed? If we
are using Solidus copying all files in the host application, why are we
still using the css/js provided by the gem? Are those files copied as
well? If yes, maybe we could just copy them in a directory that doesn't
require adding anything to the manifest?