Update test to associate a user with the order
These tests jumped immediately to an order in a checkout stage where
it is expected to have an associated user. After removing the method
`associate_user` from the before actions, it no longer adds a user when
order.user is nil. This is not an issue, but a testing error that
assumes someone could jump to an order at this stage in checkout without
already having a user assigned by the hook after_set_user (see
https://github.com/solidusio/solidus_auth_devise/blob/09841c164e1d0e6ab7afbb80494737fc692eb806/config/initializers/warden.rb#L4
).
Therefore we just need to update the test to represent what would occur
during a normal checkout flow.