Allow creating the layout with importmap support
When creating the main host app with `rails new`, we are used to
do that with the `--skip-javascript` flag. This flag was essentially
skilling all the webpack stuff that we were not supporting.

Now that Rails defaults to the old asset pipeline in combination with
importmaps, there's really no reason to keep building applications
with that flag, and we should start supporting (and using?) this tools.

When the application is made with JS, our layout will have the directive
that allows to use the importmaps, as any other Rails application.