Autocomplete main search with products and taxons
This feature uses stimulus to submit the form on every
key pressed in the search input (if > 2 chars) and add
some UX to navigate the results.

Turbo stream will take care of updating the results
section with the suggestions retrieved from the new
specific controller action introduced.

The search box changed a bit for this change. Other than the
stimulus attributes now required, the taxons filter,
which was already hidden, has been removed from the search
form. This required to change the cache specs, because that
taxons list is not generating a new cache rewrite anymore.