Remove solidus branch default value
In the test command we used to have this default, but
it's useless because we are always passing it when invoking
the command in the run-specs job (see line 86).

Having the default version in a single place will allow us
to change it only once when we create a new version branch
and we need to make its test suite point to a specific
solidus version.