Fix checkout step on CircleCI
We're bumping the browser-tools-orb version to fetch the fix
implemented at https://github.com/CircleCI-Public/browser-tools-orb/pull/63.

The issue was due to chromedriver adding a file within the project
directory, making it impossible to checkout our code there in a
following step.

To be on the safe side for future possible regressions, we're moving our
checkout step to the top of the stack to ensure it finds a pristine
directory.

Closes #315