Fix clicking on autocomplete results
Fixes #313

By clicking on one of the results item, the focusout event
is triggered and it acts before the click event is triggered,
removing the list.

To solve this, we listen to all click events and do nothing if
the click is internal to the autcomplete, otherwise it resets it.