List existing and new payment methods together
The triple level of radio buttons that was previously used was very
confusing and broken. Instead of trying to fix it seems better to
present the customer a single list of options for making the payment.

The list gives precedence to wallet payments as they require less
clicks.

If no wallet payments are available the first payment method is
selected.

Whenever a radio is selected elements from the other list are
deselected, giving the illusion of having a single list.

Each existing and new payment method have clear and distinct areas
for the input label and for the details. This gives more symmetry
between existing and new payments partial implementations and a
cleaner visual layout.